+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Deleted
+- Populations of data with python script in api entrypoint (dumps are used at db container init instead)
+
+### Added
+- Comment table
+
 ## [0.0.1] 2021-10-08
 
 ### Added